+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Improvements
+
+- Implemented `Send` and `Sync` for `GicV3`.
+
 ## 0.1.0
 
 Initial version, with basic support for GICv3 (and 4) on aarch64.